+++ packages/graph/src/utils/getDataGroupByEncodings.ts
@@ -1,4 +1,5 @@
 import { values, groupBy } from 'lodash-es';
+
 import { Encoding } from '../common/types';
 
 /**